Rename config.mjs to config.js
